pyproject.toml
[tool.poetry]
name = "swarm-manager"
version = "0.1.0"
description = ""
authors = ["Vitor Daniel <vitor036daniel@gmail.com>"]
readme = "README.md"
packages = [{include = "swarm_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
aiofiles = "^23.1.0"
aiodocker = "^0.21.0"
prisma = "^0.9.1"
python-slugify = "^8.0.1"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.22.0"}
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
